+++ source/gnss/gnssParametrization/gnssParametrizationStaticPositions.cpp
@@ -15,6 +15,7 @@
 
 #include "base/import.h"
 #include "config/config.h"
+#include "inputOutput/system.h"
 #include "files/fileInstrument.h"
 #include "files/fileGriddedData.h"
 #include "classes/platformSelector/platformSelector.h"
@@ -97,24 +98,40 @@ void GnssParametrizationStaticPositions::init(Gnss *gnss, Parallel::Communicator
       // no net positions
       noNetPos = pos0;
       if(!fileNameNoNetPositions.empty())
-        for(UInt idRecv=0; idRecv<gnss->receivers.size(); idRecv++)
-          if(platforms.at(idRecv))
-          {
-            try
+      {
+        noNetPos = std::vector<Vector3d>(pos0.size(), Vector3d(NAN_EXPR, NAN_EXPR, NAN_EXPR));
+        Bool someDisabled = FALSE;
+        do
+        {
+          someDisabled = FALSE;
+          selectedNoNetReceivers = selectNoNetReceivers->select(gnss->times.front(), gnss->times.back(), platforms);
+          for(UInt idRecv=0; idRecv<gnss->receivers.size(); idRecv++)
+            if(selectedNoNetReceivers.at(idRecv) && std::isnan(noNetPos.at(idRecv).r()))
             {
-              VariableList fileNameVariableList;
-              fileNameVariableList.setVariable("station", gnss->receivers.at(idRecv)->name());
-              Vector3dArc arc = InstrumentFile::read(fileNameNoNetPositions(fileNameVariableList));
-              auto iter = (arc.size() == 1) ? arc.begin() : std::find_if(arc.begin(), arc.end(), [&](const Epoch &e){return e.time.isInInterval(gnss->times.front(), gnss->times.back());});
-              if(iter == arc.end())
-                throw(Exception("no position found"));
-              noNetPos.at(idRecv) = iter->vector3d;
+              try
+              {
+                VariableList fileNameVariableList;
+                fileNameVariableList.setVariable("station", gnss->receivers.at(idRecv)->name());
+                if(!System::exists(fileNameNoNetPositions(fileNameVariableList)))
+                  throw(Exception("file <"+fileNameNoNetPositions(fileNameVariableList).str()+"> not exist"));
+                const Time timesMid = 0.5*(gnss->times.front()+gnss->times.back());
+                Vector3dArc arc = InstrumentFile::read(fileNameNoNetPositions(fileNameVariableList));
+                auto iter = std::min_element(arc.begin(), arc.end(), [&](const Epoch &e1, const Epoch &e2)
+                                            {return std::fabs((e1.time-timesMid).mjd()) < std::fabs((e2.time-timesMid).mjd());});
+                if(!arc.size() || ((arc.size() > 1) && (std::fabs((iter->time-timesMid).mjd()) > 0.5*medianSampling(arc.times()).mjd())))
+                  throw(Exception("No a-priori position found"));
+                noNetPos.at(idRecv) = iter->vector3d;
+              }
+              catch(std::exception &e)
+              {
+                logWarningOnce<<gnss->receivers.at(idRecv)->name()<<": "<<e.what()<<" -> not used for computation of net translation/rotation/scale"<<Log::endl;
+                platforms.at(idRecv) = nullptr;
+                someDisabled = TRUE;
+              }
             }
-            catch(std::exception &/*e*/)
-            {
-              platforms.at(idRecv) = nullptr;
-            }
-          }
+        }
+        while(someDisabled);
+      } // if(!fileNameNoNetPositions.empty())
 
       selectedNoNetReceivers = selectNoNetReceivers->select(gnss->times.front(), gnss->times.back(), platforms);
       const UInt countStation = std::count(selectedNoNetReceivers.begin(), selectedNoNetReceivers.end(), TRUE);