+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New class:        PlotDegreeAmplitudes: degreeAmplitudesSimple.
 - New option:       GnssAntennaNormalsConstraint: gnssType selection for TEC constraint.
 - New option:       PlotAxisLabeled: majorTickSpacing, minorTickSpacing, gridLineSpacing.
+- New option:       GNSS/SLR: optional disabling of stations without precise apriori positions.
 - File format:      TideGeneratingPotential includes now degree 3 tides.
 - File format:      Each file is now readable/writable in JSON format as well.
 - File format:      Removed deprecated GnssStationInfo (replaced by Platform).